Use iterator instead of loop for particle cleanup
Similar to the previous clippy recomendation, use Rust
iterator instead of a for loop to remove particles that have
reached the end of lifetime.

Note that the article mentions that the for loop goes backwards,
but that isn't neccesary for two reasons. [1]

1. The partical vector is a FIFO, not a stack. The first particals
are the oldest, so it should scan front to back.

2. The order of scanning doesn't matter since it is all done before
the screen is redrawn. Since we don't really care about what order,
the particals are removed from the vector,  we can use vector iterator
and simplify the code to retain all particals that are not finished.

[1] https://medium.com/codex/nature-of-rust-particles-40cec0a8c25e

Signed-off-by: Marc Jones <marcj303@gmail.com>